requirements.txt
selenium
beautifulsoup4
pandas
python-dateutil
python-dotenv